modif les noms des composants
